[Intl] Fix locale validator when canonicalize is true
When canonicalize is set to true, and the value length exceeds INTL_MAX_LOCALE_LEN the validator throws an exception.

The Intl Locale::canonicalize() method returns null when the value is too long and Locales::exists() only accept non null string.

This commit allows to handle the null value as it should.